feat: Refactor data access layer by renaming api to dal and implementing dynamic user profile routing
